Vergil can now Summon Doppelganger with a press of a button + various bugfixes
This was hard as fuck. (Two alternating lerps)
Bugfixes:
-DT Out SFX now properly plays after draining to 0.
-Intertia code now properly applies only when in the action AND attacking in the air state (65538)

Had to remake all of the Dopp Depletion stuff in the game using lerps and Dopp + DT Depletion. Plus made sure the variables reset properly when EventDelete is called (so as not to keep the algorithm running after teleporting to another room with Doppel On).